flaskapp/requirements.txt
anthropic
cryptography
fastapi
fastmcp
Flask
Flask-JWT-Extended
Flask-Login
Flask-SQLAlchemy
Flask-WTF
gradio
gunicorn
jinja2
mcp[cli]
numpy
openai
psycopg2-binary
pydantic
pydantic-ai
pydantic-ai-slim[mcp]
python-dotenv
requests
sqlalchemy
streamlit
tiktoken
uv
uvicorn
voyageai
websockets
Werkzeug
